Modify sorting procedure to account for offsets
The fasta fragments have a "_"-separated offset appended to them, which
prevents us from sorting the alignment files properly. This hacky awk
script removes the offset and applies it. We should probably do it
within python instead, to make it more readable and avoid problems with
exit codes when one command of the pipe fails.