Add mem requirements based on a polynomial in the total sequence length
These should probably be a config parameter rather than magic numbers
in the code, but this is just to test things out quickly.

The basic reasoning is that all memory requirements must be upper
bounded by a quadratic in the size of the total sequence, because the
maximum number of possible alignments goes as (n choose 2), which is
O(n^2).