Fix dangerous _fixPermissions behavior, just use su instead
_fixPermissions (and wrapper.sh) recursively chowned the entire
workdir to the uid/gid of the owner of the root dir. This works under
the normal "happy path" of toil, where dockerCall is given an empty
workDir that it owns. But it's catastrophically bad in certain weird
situations, like a workdir of /tmp, or if a mount is somewhere in the
work dir. This will cause (relatively minor) data loss because the
original uid/gid will be lost for all other files in the directory.

Running the docker container as the proper uid in the first place
should fix this without being so dangerous.