KTServer revamp
Makes things a bit cleaner, and gets rid of the hacky save/load
functionality that was bolted on.

To replace the hacky save/load functionality we now use an even
hackier save/load functionality, based on the existing ktserver
snapshot function. Unfortunately the toil service API is not designed
in a way that makes passing the snapshots possible. We use
semi-private toil APIs to create an empty file, which we then populate
with the snapshot after a special signal has been sent to the DB.

KT has no explicit "save" command unless using the file-backed DB,
so we have to pick up the snapshot only after we kill the server.