Get rid of maxFlowerGroupSizeRecursion indirection
The defines are a bit confusing to interpret, plus this allows
tweaking of group size per-phase. It may make sense to reduce the
group size in reference phase, for example, so the recursion doesn't
have to wait for a whole level of the tree to be finished before
starting the next.