Make SavePrimaryDB faster
We avoid reading in the entire file by just reading the first
byte. Also, this reduces the amount of waiting between polls, so small
test runs complete noticeably faster.

As a side effect of streaming the file, we have to prevent the use of
toil caching. (readGlobalFileStream doesn't have a cache=False option.)