Warn instead of quit when hitting the coverage cap
Hit this for the first time in the backbone alignment. I'm not sure
we *should* ever be hitting this cap; it may indicate things have gone
very wrong with the preprocessing step, but it's probably best to just go
ahead with the alignment anyway rather than throw a totally
unrecoverable error.