Fiddle with macOS problems again
Docker on macOS doesn't support --net=host properly, and depending on
your DNS server, (which, naturally, depends on your physical
location), your IP/hostname can be reachable or unreachable.

So the hacky "solution" is just to always use the loopback IP on
macOS. This prevents anyone from using a macOS cluster with Cactus,
but since those are few and far between it's worth sacrificing that
functionality to fix the single-machine experience.