Fix a strange PicklingError in the "child-tree" jobs
cPickle seems to have a rare bug (?) which, in certain rare cases,
causes it to think the jobs that are passed through have a bound
instance method that needs to be pickled. The python-based pickle
library doesn't have this issue, so this is a hack to force any pickle
dumps to use the Python pickle library for the problematic jobs.