Combine jobs to make preprocessor phase more autoscaler-friendly
Toil tries to run the most expensive jobs first in order to bin-pack
more efficiently. But this causes problems in the repeatmasking phase:
there are a few jobs that do very little work, but spawn larger jobs
that do much more work. Toil always schedules the large jobs first,
which has the weird effect of leaving tens of thousands of small
jobs. Those small jobs are (correctly) considered to run very quickly
by the autoscaler, so only a small number of nodes (enough to handle
the few large jobs that have appeared) are provisioned.