Allow pre-built container, fix singularity 3 pull
Added `--containerImage` option to specify a pre-built container image
instead of downloading a new one each time. Fixes issue #54.

Re-issue `singularity pull` command without the `--size` option if it
fails the first time since the `--size` option was removed in
singularity 3, but is required for version 2.3.

Clarify help text for the `--latest` option, it pulls new copy of the
latest image, but doesn't use a local copy as stated in the help.